Move Install/Uninstall plugin logic to Core.PluginManager
